// types/movie.ts
export type Movie = {
    Title: string;
    Year: string;
    imdbID: string;
    Poster: string;
};

export type MovieDetails = {
    Title: string;
    Year: string;
    Poster: string;
    Plot: string;
    Director: string;
    Actors: string;
    Genre: string;
    Runtime: string;
    imdbRating: string;
};